requirements.txt
pandas
scikit-learn
openpyxl
matplotlib
skl2onnx 
onnx
psycopg2-binary